fix: 4 个边界问题修复\n\n1. as int/as float 优先级: PREC_IN → PREC_AS，高于 + 运算符\n   '[' + (pct*100) as int + '%' 不再解析为 as 作用于字符串\n2. clib 链式调用类型推断: semantic_type.c AST_FIELD_ACCESS 加 TYPE_CLIB\n   分支+ C→Leno 映射; codegen_expr.c 无参 clib 调用生成 OP_CLIB_CALL\n3. getVolume 精度: DLL float→double, clib 声明 f32→f64\n   setVolume(0.3)→getVolume()=0.30000001(之前 4.17e-08)\n4. if-then-else: workaround 为 if/return 模式\n\n101 passed, 0 failed\

+++ leno_module/LenoMusic/src/miniaudio_dll.c
@@ -94,13 +94,13 @@ DLLEXPORT int ma_dll_sound_is_playing(SoundHandle* h) {
     return ma_sound_is_playing(h->sound) ? 1 : 0;
 }
 
-DLLEXPORT void ma_dll_sound_set_volume(SoundHandle* h, float v) {
-    if (h && h->sound) ma_sound_set_volume(h->sound, v);
+DLLEXPORT void ma_dll_sound_set_volume(SoundHandle* h, double v) {
+    if (h && h->sound) ma_sound_set_volume(h->sound, (float)v);
 }
 
-DLLEXPORT float ma_dll_sound_get_volume(SoundHandle* h) {
-    if (!h || !h->sound) return 0.0f;
-    return ma_sound_get_volume(h->sound);
+DLLEXPORT double ma_dll_sound_get_volume(SoundHandle* h) {
+    if (!h || !h->sound) return 0.0;
+    return (double)ma_sound_get_volume(h->sound);
 }
 
 DLLEXPORT int ma_dll_sound_seek(SoundHandle* h, double sec) {
